requirements.txt
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
chromadb==0.4.22
pypdf==4.0.1
python-multipart==0.0.6
ollama==0.1.6
httpx>=0.25.2,<0.26.0
sentence-transformers>=2.7.0
openai>=1.0.0
pinecone-client>=2.2.2
qdrant-client>=1.7.3
pymilvus>=2.3.4
faiss-cpu>=1.7.4